refactor: rename internal/storage/ to internal/datastore/
Generic name carries no information about what the package owns.
The package has two coupled concerns:

  - DataDir()  resolves the portable-first writable data directory
                (datadir.go, 2.4 KB)
  - TreeStore  persists per-device canonical tree snapshots
                (tree_store.go, 17 KB — the bigger half)

Either half alone is a worse name (datadir/ misses TreeStore, which
is 87% of the code). datastore/ names the whole on-disk data layer.

Path-only rename; all type names unchanged.

- folder: internal/storage/ -> internal/datastore/
- package decl: package storage -> package datastore
                (plus package storage_test -> package datastore_test)
- import path: dhs/internal/storage -> dhs/internal/datastore
- references: storage.X -> datastore.X (word-boundary safe)

## Verification
- go build ./... clean
- go vet ./... clean
- Full go test ./... clean

Refs ADR-0020, #451.
