+++ internal/datastore/identity_roundtrip_test.go
@@ -1,4 +1,4 @@
-package storage
+package datastore
 
 import (
 	"encoding/json"
