+++ internal/datastore/tree_store_test.go
@@ -1,4 +1,4 @@
-package storage
+package datastore
 
 import (
 	"os"
